% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{chrY_0_M_names_v2}
\alias{chrY_0_M_names_v2}
\title{chrY, chr0 and chrM names on v2}
\format{
An object of class \code{character} of length 7582.
}
\usage{
data(chrY_0_M_names_v2)
}
\description{
CpGs on chrY, chr0 (not mapped), chrM
}
\keyword{datasets}
