% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{generate HTML Report}
\usage{
generateReport(
  report = reportDir,
  array = arrayInput,
  pheno = NULL,
  beta_merged = betaInput
)
}
\arguments{
\item{report}{report directory}

\item{array}{array version (EPIC, mouse, v2, or other)}

\item{pheno}{pheno file (if exists)}

\item{beta_merged}{beta file}
}
\value{
html report
}
\description{
generate HTML Report
}
