% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installBiocDependencies.R
\name{installBiocDependencies}
\alias{installBiocDependencies}
\title{installBiocDependencies}
\usage{
installBiocDependencies(packageList)
}
\arguments{
\item{packageList}{packages to be installed}
}
\value{
Installs missing packages
}
\description{
Helper function to install packages when required. List of packages is contained in packageList data
}
