% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{mapping_compatible_v1v2}
\alias{mapping_compatible_v1v2}
\title{mapping IDs for probe sequence matches on EPIC v1 and v2}
\format{
An object of class \code{data.frame} with 724263 rows and 2 columns.
}
\usage{
data(mapping_compatible_v1v2)
}
\description{
mapping created using official EPICv2 and v1 manifests, latest March22, 2023
}
\keyword{datasets}
