% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{non_CpG_names}
\alias{non_CpG_names}
\title{non_CpG_names}
\format{
An object of class \code{character} of length 2932.
}
\usage{
data(non_CpG_names)
}
\description{
List of probes identified as non-CpGs on the EPIC array
}
\keyword{datasets}
