% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{non_CpG_names_v2}
\alias{non_CpG_names_v2}
\title{non_CpG_names on EPIC v2}
\format{
An object of class \code{character} of length 4439.
}
\usage{
data(non_CpG_names_v2)
}
\description{
List of probes identified as non-CpGs on the EPIC array
}
\keyword{datasets}
