% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{packageList}
\alias{packageList}
\title{packageList}
\format{
An object of class \code{character} of length 22.
}
\usage{
used in installBiocDependencies
}
\description{
required BiocDependencies
}
\keyword{datasets}
