% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocessData}
\alias{preprocessData}
\title{preprocessData}
\usage{
preprocessData(
  input = "",
  output = "",
  report = "",
  array = "EPIC",
  pheno = NULL,
  path_to_bad_sample_list = ""
)
}
\arguments{
\item{input}{Path to the input directory (raw IDAT folder)}

\item{output}{Path to output directory (this is where beta file will be saved)}

\item{report}{Path to report directory (this is where the report will be saved to)}

\item{array}{Methylation array type, EPIC by default}

\item{pheno}{Phenotypic file. NULL by default. If pheno file is provided, it must have a column named basename with identical names to IDAT files. File can be provided as Rdata, RDS, .txt or .csv. Each column will be included in the batch effect correction. Ensure that no column exists with entirely identical variables.}
}
\value{
preprocessed beta matrix and QC report
}
\description{
Pipeline to load and process raw data from the Illumina array
Authors: James E. Barrett, Chiara Herzog
Contact: james.barrett@ucl.ac.uk, chiara.herzog@uibk.ac.at
note to fix? "ChaMP_normalization/" empty folder is left in location of script where preprocessData function is called from
}
