% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessDataSigMatrix.R
\name{preprocessDataSigMatrix}
\alias{preprocessDataSigMatrix}
\title{preprocessData}
\usage{
preprocessDataSigMatrix(
  input = "",
  output = "",
  report = "",
  array = "EPIC",
  pheno = NULL,
  cores = 4,
  overwrite = FALSE,
  run.name = NULL,
  beta.subset.compatible = F,
  Uname = "Signal_A",
  Mname = "Signal_B",
  detPname = "Detection.Pval",
  sep = "\\t",
  filter.loci = F
)
}
\arguments{
\item{input}{Path to the input signal matrix}

\item{output}{Path to output directory (this is where beta file will be saved)}

\item{report}{Path to report directory (this is where the report will be saved to)}

\item{array}{Methylation array type, EPIC by default}

\item{pheno}{phenotypic file. NULL by default, should be provided}

\item{cores}{cores to be used for ChAMP normalisation. 4 by default}

\item{overwrite}{overwrite existing output folder. FALSE by default to prevent any accidental overwriting.}

\item{run.name}{NULL by default}

\item{beta.subset.compatible}{EPICv2 only, should EPIC v1/V2 compatible subset be provided?}

\item{Uname}{column name (grep) of unmethylated signal}

\item{Mname}{column name (grep) of methylated signal}

\item{detPname}{detP name (grep)}

\item{sep}{separator for index file, tab by default}

\item{filter.loci}{should detP/other sites be filtered out?}
}
\value{
preprocessed beta matrix and QC report
}
\description{
Pipeline to load and process raw data from the Illumina array
Authors: Chiara Herzog
Contact: chiara.herzog@uibk.ac.at
note to fix? "ChaMP_normalization/" empty folder is left in location of script where preprocessData function is called from
}
