% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{snp_names}
\alias{snp_names}
\title{SNP Names}
\format{
An object of class \code{character} of length 6102.
}
\usage{
data(snp_names)
}
\description{
List of probes identified as potential SNPs.
}
\keyword{datasets}
