% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{snp_names_v2}
\alias{snp_names_v2}
\title{SNP Names, remaining probes on v2}
\format{
An object of class \code{character} of length 1521.
}
\usage{
data(snp_names_v2)
}
\description{
List of probes identified as potential SNPs.
}
\keyword{datasets}
