% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_EPICversionshared_CpGs.R
\name{subset_versionshared_CpGs}
\alias{subset_versionshared_CpGs}
\title{subset_versionshared_CpGs}
\usage{
subset_versionshared_CpGs(beta, array)
}
\arguments{
\item{beta}{The preprocessed beta-matrix}

\item{array}{The type of array that is preprocessed, e.g. EPICv2}
}
\value{
A beta-matrix subset, subsetted for sequence matched cg probes both on EPIC v1 and v2, with replicated probes averaged and v1 CpG names
}
\description{
Subset a beta matrix for v1.v2 compatible CpG-probes
}
