% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{zhou_list}
\alias{zhou_list}
\title{Zhou List}
\format{
An object of class \code{character} of length 82108.
}
\usage{
data(zhou_list)
}
\description{
List of probes provided as unreliable by Zhou et al., 2017.
}
\references{
\href{https://doi.org/10.1093/nar/gky691}{Zhou et al. (2017)} Nucleic Acids Research, Volume 46, Issue 20.
}
\keyword{datasets}
