% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anno.R
\docType{data}
\name{zhou_list_v2}
\alias{zhou_list_v2}
\title{Zhou List, remaining probes on EPIC v2}
\format{
An object of class \code{character} of length 5605.
}
\usage{
data(zhou_list_v2)
}
\description{
List of probes provided as unreliable by Zhou et al., 2017.
}
\keyword{datasets}
