% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocessData}
\alias{preprocessData}
\title{preprocessData}
\usage{
preprocessData(
  input = "",
  output = "",
  report = "",
  array = "EPIC",
  pheno = NULL,
  cores = 4,
  by.dir = FALSE,
  path_to_bad_sample_list = "",
  overwrite = FALSE,
  save.rs = FALSE,
  create.shiny = F,
  find.files = F,
  run.name = NULL,
  beta.subset.compatible = F
)
}
\arguments{
\item{input}{Path to the input directory (raw IDAT folder)}

\item{output}{Path to output directory (this is where beta file will be saved)}

\item{report}{Path to report directory (this is where the report will be saved to)}

\item{array}{Methylation array type, EPIC by default}

\item{pheno}{phenotypic file. NULL by default. If pheno file is provided, it must have a column named basename with identical names to IDAT files. File can be provided as Rdata, RDS, .txt or .csv. Each column will be included in the batch effect correction. Ensure that no column exists with entirely identical variables}

\item{cores}{cores to be used for ChAMP normalisation. 4 by default}

\item{by.dir}{process by directory instead. FALSE by default. The result will not be different, but by.dir = T can be slower (for smaller projects), yet it is recommended for large projects.}

\item{path_to_bad_sample_list}{If any samples should be excluded that have previously been identified as problematic, these can be provided here as a link to a csv file}

\item{overwrite}{overwrite existing output folder. FALSE by default to prevent any accidental overwriting.}

\item{save.rs}{save SNP (rs) probe values. FALSE by default.}

\item{create.shiny}{should a shiny object for control/density visualisation be created? F by default, because this generates large objects.}

\item{find.files}{only selects those files specified in basename from a given folder. REQUIRES a pheno file with basename column, run.name, and sets by.dir to F (not applicable)}

\item{run.name}{NULL by default, required for find.files (not using plate/batch names)}

\item{beta.subset.compatible}{EPICv2 only, should EPIC v1/V2 compatible subset be provided?}
}
\value{
preprocessed beta matrix and QC report
}
\description{
Pipeline to load and process raw data from the Illumina array
Authors: James E. Barrett, Chiara Herzog, Charlotte Vavourakis
Contact: james.barrett@ucl.ac.uk, chiara.herzog@uibk.ac.at
note to fix? "ChaMP_normalization/" empty folder is left in location of script where preprocessData function is called from
}
