% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoEstimator.R
\name{rhoEstimator}
\alias{rhoEstimator}
\title{rhoEstimator}
\usage{
rhoEstimator(
  path = NULL,
  beta_e = 0.3309135,
  beta_s1 = 0.01957,
  beta_s2 = 0.97065,
  anno = "EPIC"
)
}
\arguments{
\item{path}{path to IDATs}

\item{anno}{EPIC}
}
\value{
signal-to-noise (rho) value
}
\description{
Estimates signal to noise ratio
}
