% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_EPICversionshared_CpGs.R
\name{subset_versionshared_CpGs}
\alias{subset_versionshared_CpGs}
\title{subset_versionshared_CpGs}
\usage{
subset_versionshared_CpGs(beta, array)
}
\arguments{
\item{matrix}{beta_merged}

\item{character}{"EPIC or "EPICv2"}
}
\value{
beta matrix subset
}
\description{
Subset a beta matrix for v1.v2 compatible CpG-probes
}
